backend/pyproject.toml
[tool.poetry]
name = "accident-reconstruction"
version = "0.0.1"
description = ""
authors = ["baul <baul.lee@42dot.ai>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
black = "^22.8.0"
python-dotenv = "^0.21.0"
gunicorn = "^20.1.0"
psycopg2 = "^2.9.3"
boto3 = "^1.26.3"
pandas = "^1.5.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
